[refactor] review Postman CRUD 테스트 완료

+++ bonvoyageW/src/main/java/com/bonvoyage/domain/review/dto/ReviewDto.java
@@ -17,7 +17,7 @@ public class ReviewDto {
     private String reviewContent;
     private Date writeDate;
     private String location;
-    private int userId;
-    private int contentId;
+    private Long userId;
+    private Long contentId;
 
 }
